1.15.2

+++ net/minecraft/block/MyceliumBlock.java
@@ -16,9 +16,9 @@ public class MyceliumBlock extends SpreadableBlock {
 		if (random.nextInt(10) == 0) {
 			world.addParticle(
 				ParticleTypes.field_11219,
-				(double)((float)blockPos.getX() + random.nextFloat()),
-				(double)((float)blockPos.getY() + 1.1F),
-				(double)((float)blockPos.getZ() + random.nextFloat()),
+				(double)blockPos.getX() + (double)random.nextFloat(),
+				(double)blockPos.getY() + 1.1,
+				(double)blockPos.getZ() + (double)random.nextFloat(),
 				0.0,
 				0.0,
 				0.0
